requirements.txt
Django==2.0
pkg-resources==0.0.0
pytz==2018.4
djangorestframework==3.8.2
